+++ examples/server-side-rendering/src/client/main-web.js
@@ -1,29 +1,12 @@
 import 'core-js'
 import React from 'react'
 import { hydrate } from 'react-dom'
+import { loadableReady } from '@loadable/component'
 
-import { performance } from 'perf_hooks'
+import './cypress_hooks'
 
 // eslint-disable-next-line import/no-extraneous-dependencies
-import { loadableReady, loadableEvents } from '@loadable/component'
 import App from './App'
-
-const asyncChunkLoadTimeMap = new Map()
-window.loadableEvents = function (event) {
-}
-const loadableEventsCallback = function (event) {
-  if (event.type === 'startAsyncLoad') {
-    asyncChunkLoadTimeMap.set(event.chunkName, performance.now())
-  }
-  if (event.type === 'successAsyncLoad') {
-    const t0 = asyncChunkLoadTimeMap.get(event.chunkName)
-    if (t0) {
-      event.time = performance.now() - t0
-    }
-  }
-  window.loadableEvents(event)
-}
-loadableEvents.registerCallback(loadableEventsCallback)
 
 const start = () => {
   loadableReady(() => {
